第三天

+++ src/file/DeleteDemo.java
@@ -10,11 +10,22 @@ public class DeleteDemo {
          * remove 删除
          */
         File file = new File("./test.txt");
-        boolean deleted = file.delete();
-        if (deleted){
-            System.out.println("已经删除");
+//        boolean deleted = file.delete();
+//        if (deleted){
+//            System.out.println("已经删除");
+//        }else {
+//            System.out.println("删除失败!");
+//        }
+        /*
+         * exists 存在
+         */
+        //boolean e = file.exists();
+        //System.out.println("文件存在:"+e);
+        if(file.exists()){
+            file.delete();
+            System.out.println("删除文件");
         }else {
-            System.out.println("删除失败!");
+            System.out.println("文件不存在!");
         }
     }
 }
